+++ examples/asciibird/src/main.rs
@@ -6,7 +6,7 @@ use hewn::cursor;
 use hewn::game_object::Coordinate;
 
 #[cfg(not(target_arch = "wasm32"))]
-use hewn::{display::BaseDisplay, initialize_terminal_io, io::TerminalRuntime, TerminalRenderer};
+use hewn::{display::BaseDisplay, initialize_terminal_io, runtime::TerminalRuntime, TerminalRenderer};
 fn main() {
     play_asciibird_in_terminal();
 }
