+++ examples/asciijump/src/main.rs
@@ -2,7 +2,7 @@ use asciijump::game::{SCREEN_HEIGHT, SCREEN_WIDTH};
 use hewn::cursor;
 use hewn::{game_object::Coordinate, BaseDisplay};
 #[cfg(not(target_arch = "wasm32"))]
-use hewn::{initialize_terminal_io, io::TerminalRuntime, TerminalRenderer};
+use hewn::{initialize_terminal_io, runtime::TerminalRuntime, TerminalRenderer};
 
 use crate::game::default;
