+++ src/engine.rs
@@ -4,9 +4,9 @@
 pub mod display;
 pub mod game;
 pub mod game_object;
-pub mod io;
+pub mod runtime;
 
 pub use self::display::*;
 pub use self::game_object::GameObject;
 #[cfg(not(target_arch = "wasm32"))]
-pub use self::io::initialize_terminal_io;
+pub use self::runtime::initialize_terminal_io;
