+++ src/lib.rs
@@ -30,7 +30,7 @@ pub use engine::cursor;
 pub use engine::display;
 pub use engine::game;
 pub use engine::game_object;
-pub use engine::io;
+pub use engine::runtime;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub use engine::initialize_terminal_io;
